was working now broken heart rate funcion
Code broke itself for no reason...

+++ 00_PIC/team207_main_testing.X/timers.c
@@ -2,6 +2,6 @@
 
 unsigned long millis(){
     unsigned long returnTime = 0;
-    returnTime = TMR1_ReadTimer()*scale;
+    returnTime = (unsigned long) TMR1_ReadTimer()*scale;
     return returnTime;
 }
